Add BottomTabs for Top and New deals feeds
